[Refactor] Use extract method
