+++ report.md
@@ -27,4 +27,11 @@
 
 - اعداد جادویی (۰ تا ۳۳) موجود در `switch statement` متد `semanticFunction` کلاس `CodeGenerator` را با ثابت‌های معنادار جایگزین کردم.
 
+### استفاده از الگوی `Extract Class`
+
+- با توجه به بزرگی کلاس `CodeGenerator` این کلاس برای ریفکتور کردن بسیار مناسب است. با کمی دقت متوجه می‌شویم که این کلاس کارهای بسیاری از تولید کد
+کامپایلر را برای مثال عملیات ریاضی، کنترل برنامه و... راانجام می‌دهد. با توجه به این به کمک الگوی `Extract Class` چندین
+کلاس مختلف که از نظر منطقی با آن مشابه هستند مانند `ControlFlowCodeGenerator`, `ExpressionGenerator`, `SymbolDefinitionHandler`
+را از آن استخراج می‌کنیم. البته بیشتر از این هم می‌توان از این کلاس، کلاس‌های دیگر استخراج کرد.
+
 ## سوالات
